feat: add storage persistence notice on settings page
